DDF-6626 moves ThreadContextUtils to security-core-services
