+++ platform/security/core/security-core-services/src/main/java/org/codice/ddf/security/util/ThreadContextUtils.java
@@ -11,7 +11,7 @@
  * License is distributed along with this program and can be found at
  * <http://www.gnu.org/licenses/lgpl.html>.
  */
-package org.codice.ddf.platform.util;
+package org.codice.ddf.security.util;
 
 import java.util.HashMap;
 import java.util.Map;
