+++ platform/util/platform-util/pom.xml
@@ -90,10 +90,6 @@
             <artifactId>common-system</artifactId>
             <version>${project.version}</version>
         </dependency>
-        <dependency>
-            <groupId>org.apache.shiro</groupId>
-            <artifactId>shiro-core</artifactId>
-        </dependency>
     </dependencies>
     <build>
         <plugins>
